control and view are deprecated
control and view were used in the states uit (which is also deprecated in ha 0.105)

Relevant info:

The 'control' option (with value 'hidden') is deprecated, please remove it from your configuration. This option will become invalid in version 0.107.0
The 'view' option (with value 'True') is deprecated, please remove it from your configuration. This option will become invalid in version 0.107.0


Just removed the offending lines and did a quick testrun.
Some groups might be able to be removed  (battery_view) but I'm not sure.